requirements.txt
astroid==2.3.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
future==0.17.1
-e git+https://github.com/robu3/haikoo.git@50cecd9bd38044998c741f867e0e165a4c070d32#egg=haikoo
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
line-bot-sdk==1.14.0
markovify==0.7.2
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.1.0
pylint==2.4.2
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.3
Werkzeug==0.15.6
wrapt==1.11.2
